requirements.txt
streamlit
opencv-python
ultralytics
Pillow
pandas
matplotlib
google-generativeai
gTTS
